Dissolved oxygen gages in the Klamath Basin watershed (HUC8 18010204)

Dissolved oxygen gages in the Klamath Basin watershed (HUC8 18010204)

do_gage

Format

An object of class tbl_df (inherits from tbl, data.frame) with 28 rows and 8 columns.

Examples

do_gage |> pillar::glimpse()
#> Rows: 28
#> Columns: 8
#> $ gage_name  <chr> "CDR at South Boundary", "CDR and Nutrients at Saints Rest …
#> $ gage_id    <chr> "HVTEPA_WQX-TR_SB_CDR", "HVTEPA_WQX-KR_STREST_CDR", "HVTEPA…
#> $ agency     <chr> "Hoopa Valley Tribe (Tribal)", "Hoopa Valley Tribe (Tribal)…
#> $ latitude   <dbl> 41.01731, 41.18797, 41.10679, 41.97300, 42.08459, 42.47225,…
#> $ longitude  <dbl> -123.6356, -123.6769, -123.6888, -122.3648, -122.0733, -122…
#> $ river_mile <dbl> 17.49861045, 108.83686299, 6.42925076, 6.09599928, 29.43226…
#> $ huc8       <int> 18010211, 18010209, 18010211, 18010206, 18010206, 18010203,…
#> $ stream     <chr> "Trinity River", "Klamath River", "Trinity River", "Klamath…