Publicly available flow data in the Klamath Basin watershed (HUC8 18010204)

Publicly available flow data in the Klamath Basin watershed (HUC8 18010204)

flow_data

Format

An object of class data.frame with 114654 rows and 8 columns.

Examples

flow_data |> pillar::glimpse()
#> Rows: 114,654
#> Columns: 8
#> $ stream        <chr> "Sprague River", "Scott River", "South Fork Sprague Rive…
#> $ gage_name     <chr> "SPRAGUE RIVER AT CHILOQUIN RIDGE ROAD", "Scott River Jo…
#> $ gage_id       <chr> "KLAMATHTRIBES_WQX-SR0090", "QVIR-SRJB", "KLAMATHTRIBES_…
#> $ variable_name <chr> "flow", "flow", "flow", "flow", "flow", "flow", "flow", …
#> $ value         <dbl> 763.490, 260.339, 9.250, 41.680, 92.500, 565.460, 144.68…
#> $ unit          <chr> "cfs", "cfs", "cfs", "cfs", "cfs", "cfs", "cfs", "cfs", …
#> $ statistic     <chr> "mean", "mean", "mean", "mean", "mean", "mean", "mean", …
#> $ date          <date> 2014-02-19, 2014-12-02, 2014-08-18, 2014-02-24, 2014-05…